Introduce separation between processes and the program binary
Previously, the Process class was a combination of the information
needed for a process and about the binary itself. The result was, that
restarting and similar things ended up being hacks.

The new approach is to separate the binary representation from the
actual process running the binary. Thus, the new Process class will
actually start a process on creation and run the specified binary. It is
only possible to kill it afterwards. If one now want to start a new
process, the Program class has to be used. This class contains the
information about the binary to execute and will start a new process
upon request.

While refactoring this part, also move the sampler out of the main
source file.